Random movement simpified with collision detection
