+++ src/MyCode/CollisionDetection.java
@@ -1,0 +1,13 @@
+package MyCode;
+
+import GivenCode.Field;
+import GivenCode.Location;
+
+public class CollisionDetection {
+
+    public CollisionDetection(){
+
+    }
+
+
+}
